Made Options parameter of gen() and generate() to be generic
