+++ cli2/cli.ts
@@ -400,6 +400,74 @@ async function writeFile(filePath: string, content: string) {
 }
 
 
+const generate = async (
+  options: any,
+  ir: string
+): Promise<string[]> => {
+  return new Promise((resolve, reject) => {
+    worker.ports.jsonDecodeError.subscribe((err: any) => {
+      reject(err);
+    });
+    worker.ports.generateResult.subscribe(([err, ok]: any) => {
+      if (err) {
+        reject(err);
+      } else {
+        resolve(ok);
+      }
+    });
+
+    worker.ports.generate.send([options, ir]);
+  });
+};
+
+const gen = async (
+  input: string,
+  outputPath: string,
+  options: any
+) => {
+  await fsMakeDir(outputPath, {
+    recursive: true,
+  });
+  const morphirIrJson: Buffer = await fsReadFile(path.resolve(input));
+
+  const generatedFiles: string[] = await generate(
+    options,
+    JSON.parse(morphirIrJson.toString())
+  );
+
+  const writePromises = generatedFiles.map(
+    async ([[dirPath, fileName], content]: any) => {
+      const fileDir: string = dirPath.reduce(
+        (accum: string, next: string) => path.join(accum, next), 
+        outputPath
+      );
+      const filePath: string = path.join(fileDir, fileName); 
+
+      if (await fileExist(filePath)) {
+        const existingContent: Buffer = await fsReadFile(filePath);
+
+        if (existingContent.toString() !== content) {
+          await fsWriteFile(filePath, content);
+          console.log(`UPDATE - ${filePath}`);
+        }
+      } else {
+        await fsMakeDir(fileDir, {
+          recursive: true,
+        });
+        await fsWriteFile(filePath, content);
+        console.log(`INSERT - ${filePath}`);
+      }
+    }
+  );
+  const filesToDelete = await findFilesToDelete(outputPath, generatedFiles);
+  const deletePromises = filesToDelete.map(async (fileToDelete: string) => {
+    console.log(`DELETE - ${fileToDelete}`);
+    return fs.unlinkSync(fileToDelete); 
+  });
+  return Promise.all(writePromises.concat(deletePromises));
+};
+
+
 async function writeDockerfile(
   projectDir: string,
   programOpts: any
@@ -426,4 +494,4 @@ async function writeDockerfile(
   }
 }
 
-export = { make, writeFile, fileExist, stats, writeDockerfile, findFilesToDelete, copyRedistributables };
+export = {gen,  make, writeFile, fileExist, stats, writeDockerfile, findFilesToDelete, copyRedistributables, worker };
